+++ tests/or.test.ts
@@ -1,4 +1,4 @@
-import {validatePayload, Filter} from '../src';
+import { Filter, validatePayload } from '../src/index';
 
 describe('Test or rule', () => {
   it('Passes simple test', () => {